requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
dateparser==1.0.0
DBUtils==2.0.2
Faker==9.6.0
frozenlist==1.3.0
gevent==21.8.0
greenlet==1.1.1
idna==3.2
multidict==5.1.0
mysqlclient==2.0.3
numpy==1.21.2
pandas==1.3.2
pandas-ta==0.3.14b0
pgdb==0.0.11
psutil==5.8.0
psycopg2-binary==2.9.1
PyMySQL==1.0.2
python-binance==1.0.12
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
typing-extensions==3.10.0.2
tzlocal==3.0
ujson==5.1.0
urllib3==1.26.6
websocket==0.2.1
websocket-client==1.2.1
websockets==9.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.4.0
